Small changes to Combtech - the addon should soon be updated to be equal to the one updated by Juan. This follows in next commit.
For hydrogen addon, the acronyms are updated to reflect the new technologies: SOEC and SOFC. Balbase1-3 contains now also the new equations defined for SOEC in eqndecdef. isetdecdef adds the new elements to the sets-needs to be tested! par/setdeclare and par/setdefine adds the addition of elements to sets and parameters. qeeq and qheq is affected by the SOEC technology.