Removal of superflous equations by defining properly all the equations that are supposed to be created only for investments to do so: -Definitions of IXKN, IXKN_HASORPOT, IAGKNY, IAGKN_HASORPOT were improved, and filter sets in the equations relevant (a lot of equations) were updated
VGKNACCUMNET in the equations is now correct, before the decommissioning was not 100% effective.
VGKN if fixed after each simulation for the next BB4 models to force the decommissioning
DECOM_G has been moved to the bottom of Balmorelbb4.inc to include also AGKN define in other add-ons.
OUTPUT SUMMARY has been updated (still missing to introduce the add-on stuff, e.g. hydrogen consumption)