*Introduction of three new equations to limit capacity: qktechgroupa.inc,qktechgroupr.inc and qktechgroupc.inc. They are similar to qkfuela, etc, but instead of the fuel limiting it will be the technology group. This equations should make establishing potentials more flexible and easy. *Reorganization and cleaning of Balmorelbb4.inc and Balmorelbb4.sim: For example: IAGKN, IAGKNY, IXKN etc have been taken out of the loop and restructured; and more. *Introduction of two timeseries scaling possibilities (similar to running BB3 or running BB4). I believe this makes dealing with timeseries scaling more straight forward. I will continue working on this as I believe it is still very limited and can lead to wrong results with massive penetration of renewables. *Introduction of possibility to forbid investments in generation, even though we are in BB4 mode.
Important clarification: I have been thinking a lot about the structure of the model, and I believe that it should be more flexible in term of the combinations to be used. Some of the proposed changes imply that we are moving a bit away from the BB4, BB3, BB2, BB1 traditional definitions. My proposal is to have all the elements ready, and that it should be the user deciding which combination of elements to use. This, in my opinion, makes the model super flexible in terms of the desired output (simulation of investment planning, simulation of day ahead market, etc). Example of combinations I am thinking about to be available in the near future:
Example 1: I want to run the model without investments, with the used step timeseries scaling mode, and with ....
Example 2: I want to run the model without investments, with the hydro numbers found in the preliminary analysis (something like BB3 to reduce computational time), with the used step timeseries scaling mode, and with ....
Example 3: I want to run the model with investments in generation, but only from 2030 on, with full time series linear scaling, with Unit Commitment, and with ....