Fix IM_CO2 parameter for renewable fuels
In the current version IM_CO2 accounts for the full emission from renewable fuels, while IM_CO2RE only accounts for the renewable part. IM_CO2 should only account for the fossil part. This quick fix changes that. This only had an effect on the results if FDRE in FDATA was used.